pyproject.toml
[tool.poetry]
name = "tg"
version = "0.1.0"
description = ""
authors = ["Rychanya <rychanya@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
python-telegram-bot = "^20.0"
pydantic = {extras = ["dotenv"], version = "^1.10.4"}


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
